refactor: PR #10 피드백 반영 — COUNT 타입 long, 빈 카테고리 필터 추가
- HeritageCategoryStats.total: int → long (COUNT(*) BIGINT 대응)
- HeritageMapper.xml: TRIM(category) != '' 조건 추가 (외부 API 빈 문자열 방어)

+++ src/main/java/com/histour/domain/heritage/dto/HeritageCategoryStats.java
@@ -1,3 +1,3 @@
 package com.histour.domain.heritage.dto;
 
-public record HeritageCategoryStats(String category, int total) {}
+public record HeritageCategoryStats(String category, long total) {}
